Use messaging to decouple WebUI and Comment Bot from Registration service
* Add scripts to manage services

* Move WebUI and CommentBot code to respective folders
