+++ Project.toml
@@ -8,6 +8,7 @@ AutoHashEquals = "15f4f7f2-30c1-5605-9d31-71845cf9641f"
 Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
+FileWatching = "7b1f6079-737a-58dc-b8bc-7a2ca5c1b5ee"
 GitForge = "8f6bce27-0656-5410-875b-07a5572985df"
 GitHub = "bc5e4493-9b4d-5f90-b8aa-2b2bcaad7a26"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
@@ -17,9 +18,11 @@ Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 MbedTLS = "739be429-bea8-5141-9913-cc70e7f3736d"
 Mux = "a975b10e-0019-58db-a62f-e48ff68538c9"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+Serialization = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
 Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
 TimeToLive = "37f0c46e-897f-50ef-b453-b26c3eed3d6c"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
+ZMQ = "c2297ded-f4af-51ae-bb23-16f91089e4e1"
 
 [compat]
 GitForge = ">= 0.1.1"
