+++ build.md
@@ -1,0 +1,9 @@
+# Registrator docker image
+
+Pull the latest julia image:
+
+```
+docker pull julia
+```
+
+Edit `config.toml` in `run` and make the necessary changes. Then run `build.sh`.
