--- image/README.md
@@ -1,9 +1,0 @@
-# Registrator docker image
-
-Pull the latest julia image:
-
-```
-docker pull julia
-```
-
-Copy `sample.toml` to `conf.toml` in `scripts` and make the necessary changes. Then run `build.sh`.
