--- image/scripts/.gitignore
@@ -1,1 +1,0 @@
-conf.toml
