--- image/scripts/run.jl
@@ -1,2 +1,0 @@
-using Registrator
-Registrator.RegServer.main()
