--- image/scripts/run.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-julia run.jl conf.toml > >(tee -a stdout.log) 2> >(tee -a stderr.log >&2)
