+++ run/commentbot.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env sh
+
+julia -e 'using Registrator; Registrator.CommentBot.main()' config.toml &> commentbot.log
