+++ run/config.web.toml
@@ -5,6 +5,8 @@ ip = ""
 port = 0
 server_url = ""
 registry_url = ""
+stop_file = "stopwebui"
+# log_level = "INFO"
 # registry_clone_url = ""
 # extra_providers = ""
 # registry_provider = ""
