+++ run/regservice.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env sh
+
+julia -e 'using Registrator; Registrator.RegService.main()' config.toml &> regservice.log
