+++ run/webui.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env sh
+
+julia -e 'using Registrator; Registrator.WebUI.main()' config.web.toml &> webui.log
