+++ src/Registrator.jl
@@ -7,7 +7,9 @@ import Base: PkgId
 include("slack.jl")
 include("regedit/RegEdit.jl")
 include("pull_request.jl")
-include("Server.jl")
-include("WebUI.jl")
+include("Messaging.jl")
+include("RegService.jl")
+include("commentbot/CommentBot.jl")
+include("webui/WebUI.jl")
 
 end # module
